Move integration into reference step module
At least for now, it will be too confusing to have this capability
in the ocean component's framework since it is used for verification
only.
